+++ frontend/screens/inbox/InboxInterface.scss.js
@@ -1,0 +1,274 @@
+import { Color, Padding, FontFamily, FontSize, Border } from "../GlobalStyles";
+
+export const inboxStyles = {
+    scrollViewContent: {
+        flexGrow: 1,
+      },
+      depth0Frame0Layout: {
+        backgroundColor: Color.colorWhite,
+      },
+      depth1FrameSpaceBlock: {
+        paddingHorizontal: Padding.p_base,
+        width: "100%",
+      },
+      inboxFlexBox: {
+        textAlign: "left",
+        color: Color.colorGray_500,
+      },
+      frameLayout1: {
+        width: 48,
+        height: 48,
+      },
+      frameFlexBox: {
+        justifyContent: "flex-end",
+        flexDirection: "row",
+      },
+      searchForATypo: {
+        color: Color.colorSlategray_200,
+        fontFamily: FontFamily.beVietnamProRegular,
+        textAlign: "left",
+      },
+      depth1FrameLayout: {
+        height: 72,
+        backgroundColor: Color.colorWhite,
+      },
+      frameLayout: {
+        width: 56,
+        height: 56,
+        overflow: "hidden",
+      },
+      depth3FrameSpaceBlock: {
+        marginLeft: 16,
+        height: 45,
+        justifyContent: "center",
+      },
+      searchForALayout: {
+        lineHeight: 24,
+        fontSize: FontSize.size_base,
+      },
+      depth4FrameLayout: {
+        height: 21,
+        overflow: "hidden",
+      },
+      inbox: {
+        fontSize: FontSize.size_lg,
+        letterSpacing: 0,
+        lineHeight: 23,
+        fontWeight: "700",
+        fontFamily: FontFamily.beVietnamProBold,
+      },
+      depth5Frame0: {
+        alignSelf: "stretch",
+      },
+      depth4Frame0: {
+        width: 'auto',
+        height: 23,
+        overflow: "hidden",
+        backgroundColor: Color.colorWhite,
+      },
+      depth3Frame0: {
+        width: 310,
+        height: 23,
+        alignItems: "center",
+        flexDirection: "row",
+        backgroundColor: Color.colorWhite,
+      },
+      depth5Frame01: {
+        width: 24,
+        height: 24,
+      },
+      depth4Frame01: {
+        width: 48,
+        height: 48,
+        alignItems: "center",
+        backgroundColor: Color.colorWhite,
+      },
+      depth3Frame1: {
+        flexDirection: "row",
+      },
+      depth2Frame0: {
+        alignItems: "center",
+        flexDirection: "row",
+        height: 48,
+        width: "100%",
+        backgroundColor: Color.colorWhite,
+      },
+      depth1Frame0: {
+        paddingTop: Padding.p_base,
+        paddingBottom: Padding.p_5xs,
+        height: 72,
+        backgroundColor: Color.colorWhite,
+      },
+      searchForA: {
+        lineHeight: 24,
+        fontSize: FontSize.size_base,
+      },
+      depth3Frame11: {
+        width: 'auto',
+        marginLeft: 12,
+        height: 48,
+        flex: 1, // Add flex: 1 to make it responsive
+      },
+      depth2Frame01: {
+        backgroundColor: Color.colorWhitesmoke_300,
+        borderRadius: Border.br_xs,
+        paddingVertical: Padding.p_xs,
+        alignItems: "center",
+        flexDirection: "row",
+        width: "100%",
+        paddingHorizontal: Padding.p_base,
+        height: 72,
+      },
+      depth1Frame1: {
+        height: 96,
+        paddingVertical: Padding.p_xs,
+        flexDirection: "row",
+        paddingHorizontal: Padding.p_base,
+        width: "100%",
+        backgroundColor: Color.colorWhite,
+      },
+      depth3Frame02: {
+        borderRadius: Border.br_9xl,
+        backgroundColor: Color.colorWhite,
+      },
+      projectWithAnna: {
+        fontWeight: "500",
+        fontFamily: FontFamily.beVietnamProMedium,
+        textAlign: "left",
+        color: Color.colorGray_500,
+      },
+      depth4Frame03: {
+        width: 160, //check later
+        height: 24,
+        overflow: "hidden",
+        backgroundColor: Color.colorWhite,
+      },
+      inProgress: {
+        fontSize: FontSize.size_sm,
+        lineHeight: 21,
+      },
+      depth4Frame1: {
+        width: 140,
+        backgroundColor: Color.colorWhite,
+      },
+      depth3Frame2: {
+        width: 140,
+        backgroundColor: Color.colorWhite,
+      },
+      depth2Frame02: {
+        position: "absolute",
+        top: 8,
+        left: 16,
+        height: 56,
+        alignItems: "center",
+        flexDirection: "row",
+        width: "100%",
+        backgroundColor: Color.colorWhite,
+      },
+      depth1Frame2: {
+        width: "100%",
+        backgroundColor: Color.colorWhite,
+      },
+      hiAnnaHere: {
+        fontFamily: FontFamily.beVietnamProRegular,
+        lineHeight: 24,
+        fontSize: FontSize.size_base,
+        textAlign: "left",
+        color: Color.colorGray_500,
+      },
+      depth1Frame3: {
+        height: 'auto',
+        paddingTop: Padding.p_9xs,
+        paddingBottom: Padding.p_xs,
+        backgroundColor: Color.colorWhite,
+      },
+      depth4Frame04: {
+        width: 149, //check later
+        height: 24,
+        overflow: "hidden",
+        backgroundColor: Color.colorWhite,
+      },
+      depth4Frame11: {
+        width: 129,
+        backgroundColor: Color.colorWhite,
+      },
+      depth3Frame21: {
+        width: 129,
+        backgroundColor: Color.colorWhite,
+      },
+      depth1Frame6: {
+        height: 308,
+        width: "100%",
+        backgroundColor: Color.colorWhite,
+      },
+      depth3Frame04: {
+        shadowColor: "rgba(0, 0, 0, 0.1)",
+        shadowOffset: {
+          width: 0,
+          height: 2,
+        },
+        shadowRadius: 4,
+        elevation: 4,
+        shadowOpacity: 1,
+        backgroundColor: Color.colorRoyalblue,
+        justifyContent: "center",
+        width: 56,
+        height: 56,
+        borderRadius: Border.br_xs,
+        alignItems: "center",
+        flexDirection: "row",
+        overflow: "hidden",
+      },
+      depth2Frame06: {
+        justifyContent: "flex-end",
+        flexDirection: "row",
+        backgroundColor: Color.colorWhite,
+      },
+      depth1Frame7: {
+        height: 76,
+        paddingHorizontal: Padding.p_xl,
+        paddingBottom: Padding.p_xl,
+        overflow: "hidden",
+        width: "100%",
+        backgroundColor: Color.colorWhite,
+      },
+      depth1Frame8: {
+        height: 20,
+        width: "100%",
+        backgroundColor: Color.colorWhite,
+      },
+      depth0Frame0: {
+        overflow: "hidden",
+        width: "100%",
+        backgroundColor: Color.colorWhite,
+      },
+      inboxinterface: {
+        flex: 1,
+        width: "100%",
+        backgroundColor: Color.colorWhite,
+      },
+      searchBar: {
+        flex: 1,
+        fontFamily: FontFamily.beVietnamProRegular,
+        fontSize: FontSize.size_base,
+        color: Color.colorGray_500,
+        marginLeft: 12,
+        backgroundColor: Color.colorWhitesmoke_300,
+      },
+      addButton: {
+        position: 'absolute',
+        top: 10,
+        right: 20,
+        width: 56,
+        height: 56,
+        backgroundColor: 'white',
+        borderRadius: 28,
+        alignItems: 'center',
+        justifyContent: 'center',
+        elevation: 4,
+      },
+      addButtonText: {
+        fontSize: 40,
+        color: 'Black',
+      },
+}
